Documentation, Cleanup, Adding Names to JS D3-Tree
A lot of work has been done in cleaning up the code, starting to add in unit tests into the code base, starting to clean up code structure. Still need some additional work on the REACT front end to get the names to display and some req specification as well.